README: use a single README file
The README file was duplicated in the root and profile directories. The
content was the same in both files. This commit removes the root README
file and updates the profile README file to include the content of the
root README file.

We use a sysmlink to keep the README file in the root directory.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>